Use MongoClient if blank replica set is provided
The replica set name is provided with the MONGO_REPLICA_SET environment
variable. If this is set to an empty string the MongoClient will be used
rather than the MongoReplicaSetClient.

For example:
MONGO_REPLICA_SET='' ./run_tests.sh

This is required for running in environments that do not provide a
replica set, such as travis.