Introduce mongo storage engine
This commit introduces the `MongoStorageEngine`. This is the start of a
major refactor to separate our the database access layer behind a small
strict interface that is entirely dedicated to data sets.

This commit implements `alive` as this is the simplest method.

All mongo code lives in the `mongo` module inside the
`backdrop.core.storage` package. The goal is to allow multiple storage
engines, each within their own package inside the `storage` package.