Integrate last updated feature of mongo engine
This introduces a new data-set class called NewDataSet. This is a
temporary class for the duration of the refactor that, once complete can
be renamed to DataSet. The external interface of is_recent_enough stays
exactly the same, but internally it defers to the storage engine for to get
the last update time.

This also integrates is_recent_enough with the read API detailed status view
that makes use of it.