Integrate storage engine save with write API
Integrate the new MongoStorageEngine save method with the write API and
admin app. The integration with the admin app is not right, but fixing
that at the same time is a bit too much.

This required a bit of refactoring.
1. Create a store method with the same interface as parse_and_store on
   NewDataSet
2. Refactor pre-processing steps to be a series of functions applied to
   the records and refactor tests
3. Add some tests around those steps being called
4. Integrate NewDataSet with the write API

This commit does not include removing the legacy saving code and tests.