Move Query class to core
Move the Query class from backdrop.read.query to backdrop.core.query.

The core of backdrop depends on this class so it should not be in one of
the app packages. I have left the functions used to parse a Query object
from a flask request in backdrop.read.query as they're specific to the
frontend.