Passing test for signed in user
This fixes it, tried directly finding the element and using webdrivers
text method first but this has the same incorrect idea about the dom.
Not found anything obvious about this issue in phantom 1.9.7 and don't
want to spend too long so using this hack for now. These tests may
become defunct soon anyway.