Refactor Backdrop to use the client's admin API
The performanceplatform-client now has the ability to interact with
Stagecraft, so replace the concept of DataSetRepository and
UserHttpRepository with that.

Also remove the concept of DataSetConfig and UserConfig because
those namedtuples don't really add anything anymore.

This changes the authentication mechanism from a token that Stagecraft
knows to an OAuth access token for an API user from Signon.