Remove create data set route
The mongo collections are created on the fly when written to now. Delete
has been left in for now so that a the capped size of a data set can be
changed. When a data set is removed in stagecraft it is not removed in
backdrop. This is fine for most cases as they data set can be emptied.
However, if the capped size of the data set changes this will not be
reflected in the data set unless it is first removed.