Add striped rate-limiting
Slowness for a single data-set should not result in resource-exhaustion
for the other data sets. This is a resilience measure.

Allow up to 100 requests per minute, with bursting to 5 requests per
second.

We are doing this in the application since we have the knowledge about
URLs in here, rather than in Varnish or nginx.

We need to set rate-limiting high for tests, otherwise the tests fail as
the rate-limiting kicks in.