Don’t build housekeeping branches in travis
Minor change to try to be a good travis citizen and avoid unnecessary
builds.

The deployed-to-{env} branches are used by us internally to compare what
we’re about to deploy. Those versions of the code will have already been
built in travis when we merged the pull request from a topic branch.