Fix shifting of flat data queries on boundaries
If there is a query parameter `start_at`, the Query will have a positive
delta. If `end_at` is used, the Query will have a negative delta.

In first_nonempty(), when we have a negative delta we get a first
nonempty index which is negative. The grouped version of this class
searched for the closest point across all of the groups ignoring
direction by using abs() as a parameter to min() -- which means that
min() sorts by running values through abs(), but returns the original
non-absolute value.

By including an abs() in PeriodFlatData, we were losing information
indicating the direction of shifting the data.

This bug manifested itself by providing an end_at query parameter
and a duration across a boundary where data ended, and resulted in
no data being returned for that query.