Use transformer config for tasks
Although there's an implicit dependency on backdrop.write's config (they
must be configured to use the same AMQP URL), it's too fragile to
depend on config deployed by a different fabric task.

We specify config both in tasks and the worker so that they can be more
easily split in future.