Pass date range of writes to transforms
In order for a transform to request the data that it is going to operate
over we need to know the data bounds.

This is fed through to the transform that will then handle the retrieval
of data.

All data written into the platform is required to have _timestamp so our
feature tests should reflect this. The feature tests now receive the standard
schema that any dataset will be sent by stagecraft.