Get transforms and spin off a task for each
This commit moves the tasks module to dispatch as it was clashing with
that tasks namespace.

We want to get all transforms that need running for this data set and
spin them out to a new task that will grab the data, run the transform
and write the new data.