Request data using transform configuration
We want to use the data set configuration and the query parameters
associated with the transform to request data to be transformed. We want
to set the start_at and end_at parameters using the bounding dates of
the data we want to transform so that we only work on the data that has
been changed. We make flatten explicit as we always want to be working
with this type of groupped data and it is still not the default.