Required fields for transforms (_id, _timestamp)
Backdrop requires that incoming data points should have a _timestamp and
an _id will be generated if there isn't one. Given that we want to the
transforms to be idempotent and the transforms know about the data that
is going to be submitted I have moved _id generation into them.

_timestamp is set to the start of the period in both cases, which
reflects how the front end considers the _start_at field (as a synonym
for _timestamp)