Improve transform date range handling from periods
When a collector is run it will generally POST a single record into
backdrop. This will mean that the earliest and latest values sent
through to the transform will be identical.

When working with data that is aggregated data using a `period` query
parameters using the timestamp submitted as a `start_at` and `end_at`
will result in no data. We need to using the timestamp as a `start_at`
and use a duration of one with the given period. This will be used by
backdrop to find the week that this point is a part of and return the
data for the week.