Add inclusive flag to backdrop read query
We want to be able to make a query with the end_at parameter acting
inclusively rather than exclusively. This is for the case of a single
record being added and then transformed. We pass through the earliest
and latest dates of the data being written to the transform, when there
is a single point this will be one timestamp that we cannot use to
request the data due to the exclusivity of end_at.

This results in using the `$lte` comparator in the mongo query rather
than the `$lt` comparator.