Add filter_by_prefix parameter for regex support
Add new filter_by_prefix parameter. This is similar to the existing
filter_by parameter, except that searches are anchored to the start of
the string.

The Mongo query is also constructed using 'find', but passed through as
{ 'pagePath': /^\\/housing-benefit.*/ } rather than { 'pagePath':
'/housing-benefit' }. Mongo will parse this as a regular expression
query.

We escape all strings passed into the filter using Python's re module.

The justification for this parameter is to support multi-part formats in
the info-frontend application - e.g. we would like to serve aggregate
statistics for the number of tickets submitted across all pages in
multi-part guides and smart answers. To do this with the existing
filter_by parameter would require a call to Backdrop for each page in
the guide or smart answer, and some of these formats have 30+ pages.