Refactoring and ensure we do no overwrite newer data
Move is_latest_data out to util, use in tx transform

As we check the output rather than source data set for newer data we
will never overwrite never data. This is more reliable for things with multiple potential
sources like digital takeup.

We currently test is_latest_data through the transforms using it. Arguably the stuff around calls need to be moved to
test_utils so we are not repeating tests. It is worth keeping some component integration tests for sanity though.