Add some decorators I have used for profiling
There is one decorator that will dump the arguments into and return out
of a function so that you can build a test that will isolate the
function for refactoring.

The second will dump a cProfile dump out to /tmp for the execution of a
function. This is best used wrapped around a flask route.

I have generally used `gprof2dot` to convert the profile dump into a
picture of where time is spent.