Filter by id - there is > 1 data point per slug
Previous filtering by dasboard slug means it will not post if there is
ANY newer data point for that dashboard. This made sense when we posted
all data from a time regardless of nulls and so the latest timestamp of
one data point was representative of the latest data point for all the
others. This is no longer true as we now post the latest non null for a
dashboard slug.