Ensure user sat records for incomplete periods not written
Previously the transrom written daily by the update to customer
satisfaction datsets would cause an update to the latest period daily
until the period is complete. This mean periods without enough data for
a whole week being displayed on graphs and skewing this metric by
showing data supposedly for a week in the future.

We use end at to determine completeness as if data becomes more granular
in the future then the maximum _count may change.