Bound period when getting data to transform
Given a start and end date that are both in the middle of a period and
today is also in the same period as the end date. The means that the
ending period is not fully complete in terms of data collection and we
can't accurately transform the data for it, claiming that it is
representative of the entire period.

Each of the blocks in the diagrams beneath represent a week period with
`start` and `end` pointing at the values for each of these dates.
`t` (today) points at the datetime value for now().

So:

```
  -------------------------
  |       |       |   t   |
  -------------------------
     ^              ^
   start           end
```

Becomes:

```
  -------------------------
  |       |       |       |
  -------------------------
  ^               ^
start            end
```

In the case that todays date is in the period after the one containing
the end date we want to move the end date to the end of its period. As
we should have all the data for that given period we should be able to
accurately transform its data.

So:

```
  -------------------------
  |       |       |   t   |
  -------------------------
     ^        ^
   start     end
```

Becomes:

```
  -------------------------
  |       |       |       |
  -------------------------
  ^               ^
start            end
```