Do not copy all properties from input to output
A lot of the properties that are received when querying for the input
data set are either not wanted or will break the schema validation when
POSTd as they are not meant to be part of the body.

We would not want a derivative data set to share the same auto_ids as it
will have different structure and the transforms should be generating
their own specific `_id` fields.

We only really want the following fields to be the same:
  - token: as someone might expect to be able to fiddle with derivative
    data sets
  - realtime: if the incoming data is realtime, the transformed data
    will be. This also relates to `capped_size` which should be
    configured to tell mongo how much of the realtime data to keep
  - max_age_expected: the transformed data should be as up to date as
    the input data set, so if there are rules of how old it should be
    then that should mirror to the derivative
  - published: if the input is published we would want the derivative to
    be too

`capped_size` is a little different than the others as Stagecraft will
reject a data set with this set to None, as if the key appears in the
POSTd JSON it needs to be a string. We are only interested if it is set
to a value so it gets ignored otherwise.