Allow additionalFields to be added to data
The collectors allow you to merge static data into every record sent
through to Backdrop, which is very useful for flagging up different
sources of data in a single data set.

We gain similar value from the application of tags in transforms as we
want a single data set tracking the completion rate of two different
sources (old and new forms). We would like to be able to compare them
and leave them side by side in a single data set.

This has been written so that it is applied post the transformation of
the data, regardless of what type of transform was executed. This push
for commonality in utility options is something that we have learnt is
important from the drift in config and impl in the collectors