Do no ignore none in latest tx data.
Still ignoring none from other sources (in latest_dataset_value). This
may or may not be right but this is just meant as a broad strokes demo
of how sparse just the latest data is. Also, data directly from
upload/done pages etc may be rightly ignoring nones as this is the sign
of an error - someone has deliberately uploaded a none when they should
either upload a value or not upload at all.