Ensure all data from tx ex spreadsheet same period
This is a way of ensuring all data from transactions explorer is the
same period. The previous commit represents a way of ensuring that
within a service (for all data points of a service) we take the same
period but it doesn't ensure all data is the same period as this does.

This could now be refactored to reduce the transactions explorer data to
just those values which represent the latest quarter/year and then go
through them dashboard by dashboard rather than going through each metric and
then going through the records for each dashboard to pull out the data
for the given metric. This might be slightly more efficient and clearer.