wip - don't add latest auto collected if not yet released
doing metric_to_period_mappings because collected stuff does not have a
type but we need the type to work out what the timestamp should be.

perhaps it's never quarterly/yearly? maybe it's pointless to add in?

if so then maybe refactored 350cbe271f369e992ee3b72d7f440fceb86b4e50 is
  enought for pr and we just delete these transforms

Or the client doesn't use this data as it has no type? Having it in the
dataset is fine

---

THE ANSWER: it's not quarterly/yearly. It can't fit this pattern really.
There's not way we can make the latest data point map like this:

| | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | |week|
| | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | | |q|u|a|t|e|r|-|-|-|-|-|>|
|y|e|a|r|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|-|>|

As the the way our quarters are defined don't fit week start/ends
necessarily

We will stop getting digital takeup here from collects and only get it
from tx. We will keep user sat and completion but possibly change the
messaging to make clear they don't cover the same period and also we
will remove the weighted averages as it doesn't make sense to weight
things by a different period.