Format timestamp before creating id field
Slight differences in the way the _timestamp cell is formatted (in file uploads)
results in duplicated data in the dataset.

What’s happening is this: _timestamp is listed as an auto_id for
the data set. All auto_id fields and values are joined together and
then base64 encoded and stored in the _id field.

AFTER the _id field has been created for the data record, the
_timestamp is parsed and converted to utc format. This means that
the date format of the timestamp can differ from the date format of
the timestamp in the _id.

The _id is used as a unique key for the data set.
The same _timestamp with a different date format will generate a
different _id value and will added as a new row in backdrop.

Users don't just upload new data or only the data they want to change.
Users also upload data by:
1) appending a new row to the spreadsheet and then re-uploading all
of the old data they have every time, expecting that the old data will
just be overwritten.
2) Keep the latest few rows in the spreadsheet, append the new data
to that, and then re-upload.

To handle all of these cases, the following changes were made:

1.  Move converting the _timestamp to utc format to BEFORE the
base64 encoded _id is generated.
2. Get a dict of the auto_id fields and their values.
3. Search the existing dataset for a row that matches (regardless
of the value of the _id field) and delete the row if found.

Pivotal: https://www.pivotaltracker.com/story/show/102895550