Encode unicode strings in dataset
If the dataset pushed to backdrop contains any unicode characters,
a 500 error is returned to the user.

All of the top-level strings in the dataset have been encoded as utf-8
before continuing processing and validating of the dataset.

The check is limited to only the top-level strings in a dictionary
as currently all datasets pushed to backdrop are flat.