Change development config to match other environments
The spike into Services page data anomalies has uncovered that the
configuration for transform in the development differs from that in
preview, staging in production. This means that you can't just start
backdrop and curl the transform end point update a transform data set.

In all of the other environments both backdrop write and transform use
user backdrop_write and vhost /backdrop_write for the queue config.
In development, backdrop write uses user backdrop_write and vhost /backdrop_write
for the queue config, but transforms uses user transformer and vhost /transformations.

To successfully trigger a transform in development with this set up I had to create
the transformer user and /transformations vhost manually in rabbitmq.

To make life easier for future developers, the transform config has been changed
to match backdrop_write. This change won't affect how transform run in the
other environments.