Set the celery failover strategy to round-robin
We made a change to allow the broker url to accept a list of rabbitmq
nodes. When node 1 was taken down, the celery process did not connect to
the next node in the list.

The default failover strategy for kombu is round-robin, but we
believe that celery is overwriting this and setting it to none if the
failover strategy is not explictly configured in the application
config.