Make the feature tests storage agnostic
The feature tests should not depend on mongo directly as other storage
engines will be added in future.

Usage of the storage engine (specifically the `save_record` method) causes
the current time to appear in records where it didn't before. This makes
the ETag header dynamic so the assertions have been relaxed.

We've also removed the empty_data_set feature as this is difficult to
implement in a simple, single-table SQL database.