Add postgres storage engine.
Not yet wired up.

Because of the table structure we're using in postgres we can't represent an
empty dataset. Since this isn't an essential feature we're just removing
the tests for it.

We also don't think we can represent a size cap easily in pg, so we're
skipping this test too. This feature will be required, since there are
some "realtime" datasets which will grow unboundedly otherwise. We think
we can workaround this with a regular scheduled job to remove expired
data or something like that.

Most of the tests come from base classes that cover all storage engines in
an agnostic way.