Fix intermittently failing test.
At certain times in any given hour (it is believed to affect either the
second half or the bottom half of the hour) the
"checking an out-of-date data_set" feature test will fail, during the
summer. Use UTC times in the test code and it seems to succeed all the
time.